configurable attribution, considers all campaigns in window
Prior to this change the product-attribution report considered only
engagements or events that occurred in the consideration window.
After this change the report doesnt care when the engagement or
purchase happens as long as they were against a campaign that
was sent in the consideration window.

Attribution days is now configurable from the command line.